requirements.txt
amqp==2.4.2
astroid==2.1.0
autopep8==1.4.3
billiard==3.6.0.0
certifi==2018.11.29
chardet==3.0.4
colorama==0.4.1
cssselect==1.0.3
defusedxml==0.5.0
Django==2.1.7
django-cors-headers==3.0.2
django-filter==2.1.0
django-js-asset==1.2.1
django-mptt==0.9.1
djangorestframework==3.9.1
djangorestframework-jwt==1.11.0
djangorestframework-recursive==0.1.2
gunicorn==19.9.0
idna==2.8
isort==4.3.8
kombu==4.5.0
lazy-object-proxy==1.3.1
lxml==4.3.3
Markdown==3.1.1
mccabe==0.6.1
oauthlib==3.0.1
Pillow==5.4.1
psycopg2==2.7.7
py2exe==0.9.2.2
pycodestyle==2.5.0
PyJWT==1.7.1
pylint==2.2.2
python3-openid==3.1.0
pytz==2018.9
requests==2.21.0
requests-oauthlib==1.2.0
rest-social-auth==2.0.2
six==1.12.0
social-auth-app-django==3.1.0
social-auth-core==3.1.0
sqlparse==0.3.0
urllib3==1.24.1
vine==1.3.0
wrapt==1.11.1
